Fix build failures in AppLog, LogViewerViewModel, LogViewerScreen, InternalPlaybackSourceResolver, MainActivity

+++ app/src/main/java/com/chris/m3usuite/player/internal/source/InternalPlaybackSourceResolver.kt
@@ -101,7 +101,7 @@ class PlaybackSourceResolver(
             val box = store.boxFor(ObxTelegramMessage::class.java)
             val queryBuilder = box.query()
             if (fileId != null) {
-                queryBuilder.equal(ObxTelegramMessage_.fileId, fileId)
+                queryBuilder.equal(ObxTelegramMessage_.fileId, fileId.toLong())
             } else if (messageId != null) {
                 queryBuilder.equal(ObxTelegramMessage_.messageId, messageId)
                 if (chatId != null) {
